E2E test triage creation and list
