+++ test/e2e/releases_test.go
@@ -12,7 +12,7 @@ import (
 
 func TestReleasesAPI(t *testing.T) {
 	var releases api.Releases
-	err := util.SippyRequest("/api/releases", &releases)
+	err := util.SippyGet("/api/releases", &releases)
 	require.NoError(t, err)
 
 	t.Logf("found %d releases", len(releases.Releases))
@@ -21,7 +21,7 @@ func TestReleasesAPI(t *testing.T) {
 
 func TestReleaseHealth(t *testing.T) {
 	var health api.Health
-	err := util.SippyRequest("/api/health?release="+util.Release, &health)
+	err := util.SippyGet("/api/health?release="+util.Release, &health)
 	require.NoError(t, err)
 
 	assert.Greater(t, health.Indicators["bootstrap"].CurrentRuns, 0, "no bootstrap runs")
